+++ src/main/kotlin/de/passbutler/desktop/ui/Drawables.kt
@@ -3,6 +3,7 @@ package de.passbutler.desktop.ui
 object Drawables {
     val ICON_ACCOUNT_CIRCLE = VectorDrawable("M12,2C6.48,2 2,6.48 2,12s4.48,10 10,10 10,-4.48 10,-10S17.52,2 12,2zM12,5c1.66,0 3,1.34 3,3s-1.34,3 -3,3 -3,-1.34 -3,-3 1.34,-3 3,-3zM12,19.2c-2.5,0 -4.71,-1.28 -6,-3.22 0.03,-1.99 4,-3.08 6,-3.08 1.99,0 5.97,1.09 6,3.08 -1.29,1.94 -3.5,3.22 -6,3.22z", 18, 18)
     val ICON_CHECK_CIRCLE = VectorDrawable("M12,2C6.48,2 2,6.48 2,12s4.48,10 10,10 10,-4.48 10,-10S17.52,2 12,2zM10,17l-5,-5 1.41,-1.41L10,14.17l7.59,-7.59L19,8l-9,9z", 18, 18)
+    val ICON_CONTENT_COPY = VectorDrawable("M16 1H4c-1.1 0-2 .9-2 2v14h2V3h12V1zm3 4H8c-1.1 0-2 .9-2 2v14c0 1.1.9 2 2 2h11c1.1 0 2-.9 2-2V7c0-1.1-.9-2-2-2zm0 16H8V7h11v14z", 16, 18)
     val ICON_DELETE = VectorDrawable("M6 19c0 1.1.9 2 2 2h8c1.1 0 2-.9 2-2V7H6v12zM19 4h-3.5l-1-1h-5l-1 1H5v2h14V4z", 16, 18)
     val ICON_HOME = VectorDrawable("M10 20v-6h4v6h5v-8h3L12 3 2 12h3v8z", 18, 16)
     val ICON_INFO = VectorDrawable("M12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm1 15h-2v-6h2v6zm0-8h-2V7h2v2z", 18, 18)