+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [1.1.0] - UNRELEASED
 
+### Added
+- Icon button on item detail screen to quickly copy username, password or URL to clipboard without leaving the screen
+
 ### Changed
 - Show a default confirm dialog for closing a vault because it's not a dangerous action